Implement Ord, make Eq and Hash consistent
This allows sorting email addresses, e.g. in a BTreeSet.

It also makes EmailAddress adhere to the requirement on
HashMap/HashSet that Eq and Hash are consistent. Before this change,
EmailAddress was not compliant with HashMap/HashSet's statement that:
```
k1 == k2 -> hash(k1) == hash(k2)
```
because two email addresses with different cases of domain would end up
with different hash values.